test(cactus-web): Add integration tests for accordion CACTUS-206
